requirements.txt
streamlit
Pillow
transformers
fuzzywuzzy